requirements.txt
pika
python-dotenv
mcstatus
psycopg2